% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{auto_write}
\alias{auto_write}
\title{Automatic file writer}
\usage{
auto_write(x, file, ...)
}
\arguments{
\item{x}{dataframe or datatable}

\item{file}{path to write the output file}

\item{...}{Pass any file-type appropriate arguments on to their handler
functions.}
}
\value{
data frame used as input
}
\description{
Simple tab-delim wrapper to ensure normalized function parameters
}
