% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{AutoRead}
\alias{AutoRead}
\title{Deprecated, use auto_read().
Attempts to automatically recognize file type and import
 as dataframe}
\usage{
AutoRead(file, ...)
}
\arguments{
\item{file}{path to an xlsx binary file}

\item{...}{Pass any file-type appropriate arguments on to their handler
functions. e.g., readxl::read_excel(), read.delim(), read.csv()}
}
\value{
dataframe
}
\description{
Deprecated, use auto_read().
Attempts to automatically recognize file type and import
 as dataframe
}
