% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{chomp}
\alias{chomp}
\title{Remove leading and trailing whitespace characters from a string}
\usage{
chomp(x)
}
\arguments{
\item{x}{Character string or vector.}
}
\description{
does not coerce non-character vectors so you can safely use this on an
}
