% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_merging.R
\name{CollapseDF}
\alias{CollapseDF}
\title{DEPRECATED: Collapse redundant rows of a df using the Simplify function}
\usage{
CollapseDF(df, column.names, unique = F)
}
\arguments{
\item{df}{DataFrame containing column.names}

\item{column.names}{character vector of column names used for grouping rows. 
Performs a similar function as "by=" in aggreagte()}

\item{unique}{Logical Should fields containing more then one unique value be 
allowed. It still returns a delimited list of multiple 
values but throws warning from Simplify() if enforced.}
}
\value{
collapsed df
}
\description{
This function performs similar to aggregate.data.frame, but with several
conveniences. For simplicity it currently only allows grouping
by columns that exist in df by explicit column name. Collapse columns are
moved to the front of the df.
}
\examples{
df <- data.frame(
  Patient = c(1,   1,  2,  2,  3,  4),
  Age     = c(31, 31, 32, NA, 33, NA),
  Score   = c( 9, 10,  8,  8, "",  4))
CollapseDF(df, "Patient")
 #   Patient   Age   Score
 # 1       1    31   10; 9
 # 2       2    32       8
 # 3       3    33      
 # 4       4    NA       4
}
