% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-terms.R
\name{entrez_id}
\alias{entrez_id}
\title{Entrez text search with filterability}
\usage{
entrez_id(term, field = "All Fields", db = "gene",
  organism = "Homo sapiens", retmax = 1)
}
\arguments{
\item{term}{vector search string}

\item{field}{which field type should the search terms query (https://www.ncbi.nlm.nih.gov/books/NBK49540/)}

\item{db}{Database to search (gene, protein, etc)}

\item{organism}{filter for a specific organism}
}
\value{
integer vector of entrez ids
}
\description{
General purpose function which takes a disorganized list and
returns a sorted unique list of values.
}
\examples{
 entrez_id("hsa-let-7a-1")
 entrez_id(c("hsa-let-7a-1", "tp53"))
}
