% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_tools.R
\name{GetS3Table}
\alias{GetS3Table}
\title{Read tables from S3 bucket directly into df object}
\usage{
GetS3Table(s3.path, cache = F, aws.args = "", ...)
}
\arguments{
\item{s3.path}{Full character string uri of the table.}

\item{cache}{logical value as to whether local file should be 
left, default for cache=F removes file after transfer.}
}
\value{
data.frame
}
\description{
Works for all delimited tables supported by toolboxR::Autoread().
which includes csv, tab-delim, and xlsx). Only
}
