% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-terms.R
\name{gpsdb_table}
\alias{gpsdb_table}
\title{Gene and Protein Synonyms Table from ExPASy}
\usage{
gpsdb_table(term, acc = "", unique_only = TRUE)
}
\arguments{
\item{term}{The term you'd like to search for.}

\item{acc}{The corresponding UniProt accession number, used for filtering when "term" returns multiple results}

\item{unique_only}{Default = TRUE, otherwise you can return multiple results.}
}
\description{
Fetch a synonym table from the GPSDB synonym tool this tool returns synonyms
from EntrezGene, HGNC, OMIM and SwissProt(UniProt) databases if your search
term matched more than one geneset, it will be reflected in the "geneset" values
}
\examples{
gpsdb_table("npc1")
}

