% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_excel}
\alias{import_excel}
\title{Read an excel file and keep blank cells}
\usage{
import_excel(file, sheet = 1)
}
\arguments{
\item{sheet}{sheet name or number, defaults to first sheet}

\item{xlsxFile}{path to an xlsx binary file}
}
\value{
dataframe
}
\description{
Read an excel file and keep blank cells
}

