% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_df.R
\name{normalize_strings}
\alias{normalize_strings}
\title{Automated normalization of string formats}
\usage{
normalize_strings(x, case = "upper", length = 20)
}
\arguments{
\item{x}{Character vector.}

\item{case}{Character value "upper" or "lower"}

\item{length}{integer length which we will trim longer strings to. If non-alphanumeric
characters exist (e.g., dashes or spaces) we will trim +/-5 characters
of this length to retain intact words.}
}
\value{
character vector
}
\description{
I commonly use this to clean abnormally formatted df variable names.
}
