% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdrc_to_ggplot2.R
\name{plotdrc_to_ggplot2}
\alias{plotdrc_to_ggplot2}
\title{Plot a drc plot to a ggplot2 plot}
\usage{
plotdrc_to_ggplot2(x, ..., add = FALSE, level = NULL,
  type = c("average", "all", "bars", "none", "obs", "confidence"),
  broken = FALSE, bp, bcontrol = NULL, conName = NULL, axes = TRUE,
  gridsize = 100, log = "x", xtsty, xttrim = TRUE, xt = NULL,
  xtlab = NULL, xlab, xlim, yt = NULL, ytlab = NULL, ylab, ylim, cex,
  cex.axis = 1, col = FALSE, lty, pch, legend, legendText, legendPos,
  cex.legend = 1, normal = FALSE, normRef = 1,
  confidence.level = 0.95, plot_title = NULL)
}
\description{
Plot a drc plot to a ggplot2 plot
}
