% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_tools.R
\name{PutS3Table}
\alias{PutS3Table}
\title{Write tab-delim tables to S3 bucket}
\usage{
PutS3Table(object, s3.path, sep = "\\t", row.names = F, quote = F,
  aws.args = "", ...)
}
\arguments{
\item{object}{a table object (df, data.table, tibble}

\item{s3.path}{Full character string uri of the table.}
}
\value{
aws response code
}
\description{
Write tab-delim tables to S3 bucket
}
