% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/report_templates.R
\name{rancho_project_report}
\alias{rancho_project_report}
\alias{rancho_project_report}
\title{Rancho project report latex template}
\format{A latex template document}
\source{
edited from default.latex distributed with pandoc
}
\usage{
rancho_project_report()

rancho_project_report()
}
\value{
a pdf report generated from pandoc
}
\description{
Rancho project report latex template

Custom PDF format for Rancho Project Reporting
}
