% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_df.R
\name{remove_columns}
\alias{remove_columns}
\title{Sort specified columns to front of a data.frame}
\usage{
remove_columns(names, df)
}
\arguments{
\item{names}{a char vector of current columns yto remove}

\item{df}{data frame containing columns "names"}
}
\value{
A df with rearranged columns
}
\description{
Sort specified columns to front of a data.frame
}

