% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_df.R
\name{rename_columns}
\alias{rename_columns}
\title{Renames specified columns in data.frame}
\usage{
rename_columns(old, new, df)
}
\arguments{
\item{old}{current column name}

\item{new}{replacement column name}

\item{df}{which data frame to replace it in}
}
\value{
A df with changed column names
}
\description{
Renames specified columns in data.frame
}
