% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew_packages.R
\name{renew_packages}
\alias{renew_packages}
\title{This  script installs all neccessary packages after 
an update to a new version of R and/or RStudio}
\usage{
renew_packages()
}
\description{
This  script installs all neccessary packages after 
an update to a new version of R and/or RStudio
}
