% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_axis_labels.R
\name{rotate_axis_labels}
\alias{rotate_axis_labels}
\title{Rotate the x of y axis labelds of a ggplot graph}
\usage{
rotate_axis_labels(axis, angle, hjust)
}
\description{
Rotate the x of y axis labelds of a ggplot graph
}
