% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_rmd_chunks.R
\name{run_all_rmd_chunks}
\alias{run_all_rmd_chunks}
\title{A function to evaluate all chunks in an Rmd file}
\usage{
run_all_rmd_chunks(rmd, envir = globalenv())
}
\description{
A function to evaluate all chunks in an Rmd file
}
