% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_tools.R
\name{s3_set_env}
\alias{s3_set_env}
\title{Establish an environment which defines the current S3 bucket metadata.}
\usage{
s3_set_env(bucket = NULL, wd = NULL, local.cache = "/tmp/s3.cache")
}
\arguments{
\item{bucket}{base name for s3 bucket e.g. s3://bucket}

\item{wd}{character string path, optional to define a working prefix}

\item{local.cache}{character string, path of a local scratch directory}
}
\value{
environment e returned invisibly to global environment
}
\description{
This environment is required for all of my S3... functions. Sets bucket name,
defines and creates a local cache directory, tracks working directory
}
