% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_df.R
\name{select_to_front}
\alias{select_to_front}
\title{Sort specified columns to the front of a data frame}
\usage{
select_to_front(df, ..., after)
}
\arguments{
\item{df}{data.frame with existing columns}

\item{...}{Unquoted column names to place at front of df}

\item{after}{Optional unquoted column name to place
specified columns immediately after}
}
\value{
A df with rearranged columns
}
\description{
Sort specified columns to the front of a data frame
}
