% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_and_summarizing_tables.R
\name{set.value}
\alias{set.value}
\title{Generate function to set the value of a specified row/column}
\usage{
set.value(id)
}
\arguments{
\item{id}{Used during function creation to customize which df columns
are used for search. This can be a vector.}

\item{X}{Search term, this can be a vector}

\item{dat}{a data frame with both "id" and "field" columns}

\item{field}{the search column name}
}
\value{
value from field corresponding to search terms. This function
        can select multiple field values that are collapsed with
        separator to conserve dimensionality
}
\description{
A helper function to build df-specific search tools.
These functions are typically used in conjunction with lapply to
determine if they are present in a df.
}
