% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_tools.R
\name{WriteTableToS3}
\alias{WriteTableToS3}
\title{Helper function for writing tab-delimited tables to an S3 bucket}
\usage{
WriteTableToS3(bucket, path = "", arguments = "")
}
\arguments{
\item{bucket}{S3 bucket name}

\item{path}{S3 prefix structure to prepend to the name, 
leading and trailing slashes are cleaned up.}

\item{obj}{The Dataframe  you want to write}

\item{name}{File name for the table}
}
\value{
Response code from AWS, "0" when successful
}
\description{
A helper function to build S3 path-specific tools. Since you will commonly write
to a single path in an S3 bucket repeatedly, this simplifys the process by
pre-populating longer arguments such as bucket and path strings.
}

