% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_merging.R
\name{append_df}
\alias{append_df}
\title{Append DF using specified overwrite modes}
\usage{
append_df(main, new, id = "Patient", mode = "safe", verbose = TRUE,
  delim = "; ")
}
\arguments{
\item{main, new}{data frames, each with column "id".}

\item{id}{column name used to designate matching rows.}

\item{mode}{append : append any new values to preexisting values using delimiter}

\item{verbose}{not yet implemented}

\item{delim}{string value used to append new values during "append" mode.}
}
\value{
merged df
}
\description{
Given a main df, either replace, append, or warn of conflict for values
 found in matching column and row of new df. Row is matched by a paired id
 value.
}

