% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{autoread}
\alias{autoread}
\title{Attempts to automatically recognize file type and import
 as dataframe}
\usage{
autoread(file, sheet = 1)
}
\arguments{
\item{file}{path to an xlsx binary file}

\item{sheet}{number or name of workbook sheet to import.
Defaults to first sheet.}
}
\value{
dataframe
}
\description{
Attempts to automatically recognize file type and import
 as dataframe
}

