% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{AutoRead}
\alias{AutoRead}
\title{Attempts to automatically recognize file type and import
 as dataframe}
\usage{
AutoRead(file, sheet = 1, check.names = T, skip = 0, colnames = T)
}
\arguments{
\item{file}{path to an xlsx binary file}

\item{sheet}{number or name of workbook sheet to import.
Defaults to first sheet.}

\item{check.names}{logical. If TRUE then the names of the variables in the 
data frame are checked to ensure that they are syntactically 
valid variable names. If necessary they are adjusted 
(by make.names) so that they are, and also to ensure 
that there are no duplicates.}

\item{skip}{number of lines to skip before import.}
}
\value{
dataframe
}
\description{
Attempts to automatically recognize file type and import
 as dataframe
}
