% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_df.R
\name{format_column_names}
\alias{format_column_names}
\title{Format columns names to lower, UPPER or Title.Case}
\usage{
format_column_names(df, type = "title")
}
\arguments{
\item{df}{which data frame to replace it in}

\item{type}{default="title", also accepts "lower" or "upper"}
}
\value{
df with renamed columns
}
\description{
Format columns names to lower, UPPER or Title.Case
}
